requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scipy>=1.7.0
statsmodels>=0.13.0
networkx>=2.6.0
pyyaml>=6.0
PyQt5>=5.15.0
tqdm>=4.61.0
torch>=1.10.0